+++ react/ProgressBar.js
@@ -1,0 +1,3 @@
+import ProgressBar from './components/ProgressBar/index'
+
+export default ProgressBar